vegflow_backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
django-filter>=23.5
Pillow>=10.0
